+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - YYYY-MM-DD
+### ⚠ Breaking Changes
+* datamine-types: `MoveType` values are now strings instead of numbers.
+* datamine-types: `IUnitMovementEntry['move speed']` is now an optional value.
+* datamine-types: `IItem.dictionary.lore` has been replaced with `IItem.lore`.
 
 ## [0.5.0] - 2020-05-03
 ### ⚠ Breaking Changes