+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	* `IMission.assets`
 	* `UnitGettingType.ExtraSkillElgif`
 	* `IUnit.specialEffects`
+	* `IUnit.associated_elgifs` and `IItem.associated_elgifs`
 
 ## [0.5.0] - 2020-05-03
 ### ⚠ Breaking Changes