+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	* `UnitGettingType.ExtraSkillElgif`
 	* `IUnit.specialEffects`
 	* `IUnit.associated_elgifs` and `IItem.associated_elgifs`
+	* `IItem.appearanceRequirements`
 
 ## [0.5.0] - 2020-05-03
 ### ⚠ Breaking Changes