+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	* `IUnit.specialEffects`
 	* `IUnit.associated_elgifs` and `IItem.associated_elgifs`
 	* `IItem.appearanceRequirements`
+	* `IUnknownClearBonus`
 
 ## [0.5.0] - 2020-05-03
 ### ⚠ Breaking Changes