+++ CHANGELOG.md
@@ -12,8 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * datamine-types: Remove `IExtraSkillCondition` and replace with `ExtraSkillCondition` type.
 	* Split into subtypes `IExtraSkillItemCondition`, `IExtraSkillSphereTypeCondition`, `IExtraSkillUnitCondition`, `IExtraSkillUnknownCondition`.
 	* Changes type of `IExtraSkillPassiveEffect.conditions` and `IExtraSkillUnknownPassiveEffect.conditions` to be `ExtraSkillCondition[]`.
-* datamine-types: Move `effect delay time(ms)/frame` property from `IDamageFramesEntry` to `IBurstDamageFramesEntry`
-	* As a result, `getExtraAttackDamageFramesEntry` now returns `IBurstDamageFramesentry` instead of `IDamageFramesEntry`
+* datamine-types: Move `effect delay time(ms)/frame` property from `IDamageFramesEntry` to `IBurstDamageFramesEntry`.
+	* As a result, `getExtraAttackDamageFramesEntry` now returns `IBurstDamageFramesentry` instead of `IDamageFramesEntry`.
+* datamine-types: Replace `IUnit.evolution.cost` with `IUnit.evolution.zel` and `IUnit.evolution.karma`.
+* datamine-types: `IUnit.evolution.prev` is now an array of strings instead of just a string.
 
 ### ➕ Additions
 * datamine-types: Add definitions for the following types